requirements.txt
atomicwrites==1.4.1
attrs==21.4.0
colorama==0.4.5
cycler==0.11.0
fonttools==4.34.4
iniconfig==1.1.1
kiwisolver==1.4.4
matplotlib==3.5.2
numpy==1.23.1
numpy-stl==2.17.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pydicom==2.3.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-utils==3.3.3
six==1.16.0
tomli==2.0.1
